+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - Broker#run_once now returns true if a message was handled.
+- PitchBendChange.create now accepts a float value instead of the raw bytes.
 
 ## [0.2.0] - 2018-04-11
 ### Added